+++ models/User.model.js
@@ -26,4 +26,6 @@ const UserSchema = new Schema({
     }
 });
 
-module.exports = mongoose.model("User", UserSchema);
+//creating the model
+const User = mongoose.model("User", UserSchema);
+module.exports = User
